Switched to configuration summary plugin for a Roddy-associated group-versions.py script.
Now only the plugin-specific configuration logic is implemented in the plugin, while the global logic for grouping is delivered with Roddy.
